requirements.txt
django>=2.0.8
sphinx==1.7.*
django-oauth-toolkit==1.1.*
psycopg2==2.7.*
djangorestframework==3.8.*
django-rest-framework-social-oauth2==1.1.*
social-auth-app-django==2.1.*
